feat(add.tsx): create Add component to add movies to the database feat(addMovie.ts): create API endpoint to add movies to the database feat([id].tsx): create movie page to display movie details
The changes add functionality to the application to allow users to add movies to the database. The Add component is created to provide a form for users to enter the IMDB ID of the movie they want to add. The AddMovie API endpoint is created to handle the addition of movies to the database. The [id].tsx page is created to display the details of a movie.