feat(typings): add typings for movie, page, and search
Add typings for movie, page, and search to provide a clear structure for the data. The `MovieType` interface describes the structure of the data for a movie item. The `Page` type describes a page of items, including an array of items and metadata about the page. The `SearchResult` type describes the structure of the data for a search result, including an array of movies and metadata about the search.